etl/requirements.txt
pandas
google-cloud-storage
google-cloud-bigquery
python-dotenv
flask
pyarrow